+++ src/main/app.ts
@@ -126,6 +126,7 @@ export class KandoApp {
         soundVolume: 0.5,
         sidebarVisible: true,
         ignoreWriteProtectedConfigFiles: false,
+        trayIconFlavor: 'color',
         enableVersionCheck: true,
         zoomFactor: 1,
         menuOptions: {
@@ -208,6 +209,11 @@ export class KandoApp {
     this.appSettings.onChange('ignoreWriteProtectedConfigFiles', (newValue) => {
       this.appSettings.ignoreWriteProtectedConfigFiles = newValue;
       this.menuSettings.ignoreWriteProtectedConfigFiles = newValue;
+    });
+
+    // Update the tray icon if the tray icon flavor changes.
+    this.appSettings.onChange('trayIconFlavor', () => {
+      this.updateTrayMenu(true);
     });
 
     // Initialize the IPC communication to the renderer process.
@@ -987,20 +993,54 @@ export class KandoApp {
     this.bindingShortcuts = false;
   }
 
-  /** This updates the menu of the tray icon. It is called when the menu settings change. */
-  private updateTrayMenu() {
+  /**
+   * This updates the menu of the tray icon. It is called when the menu settings change or
+   * when the tray icon flavor changes.
+   */
+  private updateTrayMenu(flavorChanged = false) {
+    // If the flavor of the tray icon has changed, we have to destroy the old tray icon
+    // and create a new one.
+    if (flavorChanged) {
+      this.tray?.destroy();
+      this.tray = null;
+    }
+
+    // If the tray icon flavor is set to 'none', we do not show a tray icon.
+    let flavor = this.appSettings.get('trayIconFlavor');
+    if (flavor === 'none') {
+      return;
+    }
+
+    // The tray icons are not bundled via webpack, as the different resolutions for HiDPI
+    // displays on macOS or the flavors on Linux and Windows are loaded at runtime.
+    // Instead, the tray icons are copied to the assets directory during the build
+    // process. See webpack.plugins.ts for more information.
     if (!this.tray) {
       if (os.platform() === 'darwin') {
-        // On macOS, the tray icons are not bundled via webpack, as the different
-        // resolutions for HiDPI displays are loaded at runtime. Instead, the tray icons
-        // are copied to the assets directory during the build process.
-        // See webpack.plugins.ts for more information.
         this.tray = new Tray(path.join(__dirname, '../renderer/assets/trayTemplate.png'));
       } else {
-        this.tray = new Tray(
-          path.join(__dirname, require('../../assets/icons/icon.png'))
-        );
+        if (flavor !== 'light' && flavor !== 'dark' && flavor !== 'color') {
+          console.warn(`Unknown tray icon flavor: '${flavor}'. Using 'color' instead.`);
+          flavor = 'color';
+        }
+
+        let iconPath;
+
+        switch (flavor) {
+          case 'light':
+            iconPath = path.join(__dirname, require('../../assets/icons/tray-light.png'));
+            break;
+          case 'dark':
+            iconPath = path.join(__dirname, require('../../assets/icons/tray-dark.png'));
+            break;
+          case 'color':
+            iconPath = path.join(__dirname, require('../../assets/icons/tray-color.png'));
+            break;
+        }
+
+        this.tray = new Tray(iconPath);
       }
+
       this.tray.setToolTip('Kando');
     }